chore(effect): remove migration scaffolding and dead bridges
Delete the phased migration doc now that the work is complete, collapse
InMemorySessionRegistry onto the Effect implementation, drop deprecated
toPromiseSessionRegistry, and remove unused web Effect wrapper layers that
only re-exported Promise handlers for tests.
