+++ apps/web/package.json
@@ -29,7 +29,6 @@
     "@tailwindcss/vite": "4.3.3",
     "better-auth": "1.6.29",
     "drizzle-orm": "0.45.2",
-    "effect": "3.22.1",
     "marked": "18.0.9",
     "pg": "8.23.0",
     "react": "19.2.8",
