--- apps/web/src/effect/index.ts
@@ -1,6 +1,0 @@
-export {
-  fromPromise,
-  fromPromiseWithError,
-  runPromiseExit,
-  runPromiseOrDie,
-} from "./runtime-bridge.js";
