--- apps/web/src/effect/runtime-bridge.ts
@@ -1,6 +1,0 @@
-export {
-  fromPromise,
-  fromPromiseWithError,
-  runPromiseExit,
-  runPromiseOrDie,
-} from "@sidecar/core/effect";
